se arreglo el add/POST

+++ src/main/java/com/example/modeladov1/model/Categoria.java
@@ -27,18 +27,18 @@ public class Categoria {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_categoria")
-    private Long id_categoria;
+    private int id_categoria;
 
     @Column(name="nombre")
     private String nombre;
 
     // Getters y setters
 
-    public Long getId_categoria() {
+    public int getId_categoria() {
         return id_categoria;
     }
 
-    public void setId_categoria(Long id_categoria) {
+    public void setId_categoria(int id_categoria) {
         this.id_categoria = id_categoria;
     }
